+++ items/main.sh
@@ -34,7 +34,7 @@ static_content() {
         0.0 M W "handling Window    $nav_next" windows.sh \
         0.0 M S "handling Sessions  $nav_next" sessions.sh \
         0.0 M B "paste Buffers      $nav_next" paste_buffers.sh \
-        2.0 M M "Missing keys       $nav_next" missing_keys.sh \
+        2.0 M M "Missing keys       $nav_next" "$d_odd_chars"/missing_keys.sh \
         0.0 M A "Advanced options   $nav_next" advanced.sh \
         0.0 M E "Extras             $nav_next" extras.sh
     menu_generate_part 1 "$@"